Remove Agents class (#213)
We had this experimental class called Agents that was basically a
"stateful" ChatPrompt. To align with other languages, this class doesn't
add much value, and is taking on additional support pressure. So I'm
going to remove it in interest of just ChatPrompt.

ChatPrompt now supports "memory" so if you give it memory when you
construct it, it'll use that memory instance for subsequent calls (which
is basically what Agent did).

So TLDR: Change `Agent` to `ChatPrompt` in your code to migrate.